+++ actix-http/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.13.2
+
 - Fix `MessageBody` implementation for `ByteString` to signal completion after yielding its contents. [#4177]
 
 [#4177]: https://github.com/actix/actix-web/issues/4177